copy commented out impl to another file
rename RenderViewController with UIScrollView to
RenderViewController2
