+++ playgrounds/RenderViewController.swift
@@ -6,6 +6,8 @@
 //  Copyright © 2017年 sonson. All rights reserved.
 //
 
+//  Second Implementation with UITableView
+
 import UIKit
 
 public func addLine(x: [Double], y: [Double]) {
@@ -52,83 +54,6 @@ public func makeRenderer() -> LineGraphRenderer {
         LineData(points: DummyData.points2())
         ])
 }
-
-/*
-public class RenderViewController: UIViewController {
-
-    func makeRenderer() -> LineGraphRenderer {
-        return LineGraphRenderer(lines: [
-            LineData(points: DummyData.points1()),
-            LineData(points: DummyData.points2())
-            ]
-        )
-    }
-
-
-    static var shared:RenderViewController!
-
-    var renderers:[Renderer] = []
-
-    var scrollView: UIScrollView!
-
-    public override func viewDidLoad() {
-        super.viewDidLoad()
-
-        RenderViewController.shared = self
-
-        scrollView = UIScrollView()
-        scrollView.frame = self.view.frame
-
-        self.view.addSubview(scrollView)
-    }
-
-
-    public override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
-
-        self.render()
-
-    }
-
-
-    public override func viewDidLayoutSubviews() {
-        super.viewDidLayoutSubviews()
-        updateViews()
-    }
-
-    public override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-    }
-
-    func updateViews(){
-        for view in scrollView.subviews{
-            view.removeFromSuperview()
-        }
-        scrollView.contentSize.height = 0
-        self.render()
-    }
-
-
-    func render(){
-
-        var size = self.view.frame.size
-        size.height *= 0.5
-        // renderer 取り出してscrollviewに追加
-        for renderer in renderers {
-            let image = renderer.render(size: size)
-            let imageView = UIImageView(image: image)
-            imageView.frame.size = size
-            imageView.contentMode = .scaleToFill
-            imageView.frame.origin = CGPoint(x: 0, y: scrollView.contentSize.height)
-            scrollView.addSubview(imageView)
-            scrollView.contentSize.height += size.height
-        }
-    }
-    
-    
-}
-
- */
 
 private class RenderTableViewCell: UITableViewCell {
